Add files via upload
This academic portfolio is prepared to compile selected documents and proofs to illustrate my accomplishments during the past years. The earned experiences, knowledge and skills encouraged me to integrate my achievements to develop a new idea. The focus would be to design marketing strategies for the local shops and plan hands-on workshops for the managers to develop their own digital transformation. The aim is to improve the performance of small businesses through recommending tailored marketing strategies and educating the retailers to build digital literacy skills.